// front/src/api/model/todoListItem.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * Todo のリスト
 */
export interface TodoListItem {
  /** 完了しているか */
  completed: boolean;
  /** Todo ごとに一意な ID */
  id: string;
  /** テキスト */
  text: string;
}